# backend/src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Online Voting System API",
    "version": "1.0.0",
    "description": "Secure Online Voting System Backend APIs for authentication, voter profiles, elections, candidates, voting, notifications and administration."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server"
    },
    {
      "url": "https://learning-rbko.onrender.com",
      "description": "Production Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Election": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "65abc123456789abcdef1234"
          },
          "title": {
            "type": "string",
            "example": "Student Council Election 2026"
          },
          "description": {
            "type": "string",
            "example": "Annual student council election"
          },
          "electionType": {
            "type": "string",
            "enum": [
              "PRESIDENTIAL",
              "PARLIAMENTARY",
              "ASSEMBLY",
              "LOCAL",
              "COLLEGE",
              "ORGANIZATION",
              "OTHER"
            ],
            "example": "COLLEGE"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "UPCOMING",
              "LIVE",
              "COMPLETED",
              "CANCELLED"
            ],
            "example": "UPCOMING"
          },
          "isPublished": {
            "type": "boolean",
            "example": true
          },
          "isResultsPublished": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "Candidate": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "65abc123456789abcdef5678"
          },
          "election": {
            "type": "string",
            "example": "65abc123456789abcdef1234"
          },
          "name": {
            "type": "string",
            "example": "Rahul Sharma"
          },
          "party": {
            "type": "string",
            "example": "Student Unity Party"
          },
          "symbol": {
            "type": "string",
            "example": "⭐"
          },
          "photo": {
            "type": "string",
            "example": "https://example.com/photo.jpg"
          },
          "manifesto": {
            "type": "string"
          },
          "biography": {
            "type": "string"
          },
          "experience": {
            "type": "string"
          },
          "position": {
            "type": "string",
            "example": "President"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "Profile": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Manisha Pabolta"
          },
          "age": {
            "type": "integer",
            "minimum": 18,
            "example": 21
          },
          "gender": {
            "type": "string",
            "enum": [
              "Male",
              "Female",
              "Other"
            ],
            "example": "Female"
          },
          "address": {
            "type": "string",
            "example": "New Delhi"
          },
          "city": {
            "type": "string",
            "example": "Delhi"
          },
          "state": {
            "type": "string",
            "example": "Delhi"
          },
          "pincode": {
            "type": "string",
            "example": "110001"
          },
          "phone": {
            "type": "string",
            "example": "+919876543210"
          },
          "aadhaarNumber": {
            "type": "string",
            "example": "123456789012"
          },
          "voterId": {
            "type": "string",
            "example": "VOTER-123456"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get All Users",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved successfully"
          },
          "403": {
            "description": "Admin access required"
          }
        }
      }
    },
...
    "/api/profile": {
      "post": {
        "tags": [
          "Voter Profile"
        ],
        "summary": "Create Voter Profile",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Profile"
                  }
                ],
                "type": "object",
                "properties": {
                  "idProof": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Voter profile created successfully"
          }
        }
      },

      "put": {
        "tags": [
          "Voter Profile"
        ],
        "summary": "Update Voter Profile",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Voter profile updated successfully"
          }
        }
      }
    },
...
    "/api/elections": {
      "get": {
        "tags": [
          "Elections"
        ],
        "summary": "Get All Elections",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "All elections retrieved successfully"
          }
        }
      },

      "post": {
        "tags": [
          "Elections"
        ],
        "summary": "Create Election",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Election"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Election created successfully"
          }
        }
      }
    },
...
    "/api/candidates": {
      "get": {
        "tags": [
          "Candidates"
        ],
        "summary": "Get Candidates",
        "parameters": [
          {
            "name": "electionId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Candidates retrieved successfully"
          }
        }
      },

      "post": {
        "tags": [
          "Candidates"
        ],
        "summary": "Create Candidate",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Candidate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Candidate created successfully"
          }
        }
      }
    },
...
    "/api/notifications": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Get Notifications",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Notifications retrieved successfully"
          }
        }
      }
    },
...
  }
}